src/core/pyproject.toml
[tool.poetry]
name = "wenqu"
version = "0.1.0"
description = "生产级RAG（Retrieval Augmented Generation）框架，支持PDF/DOCX/DOC/PPTX/EXCEL/MARKDOWN/TXT等多种文件解析，以模块化方式构建RAG流程。"
authors = ["Wenbao Xu <xu-wenbao@foxmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
markdown-it-py = "^3.0.0"
mdformat = "^0.7.17"
mdformat-gfm = "^0.3.6"
mdformat-tables = "^1.0.0"
linkify-it-py = "^2.0.3"
langchain = ">=0.2,<0.3"
loguru = "^0.7.2"
more-itertools = "^10.5.0"
beautifulsoup4 = "^4.12.3"
blinker = "^1.8.2"
fast-langdetect = "^0.2.2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[tool.poetry.group.dev.dependencies]
ragas = "^0.1.21"
jupyterlab = "^4.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
